requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
async-timeout==4.0.2
attrs==21.2.0
Automat==20.2.0
Babel==2.8.0
bcrypt==3.2.0
blinker==1.4
certifi==2020.6.20
chardet==3.0.4
charset-normalizer==3.2.0
click==8.0.3
cloudpickle==2.2.1
colorama==0.4.4
configobj==5.0.6
cryptography==3.4.8
dbus-python==1.2.18
distro==1.7.0
frozenlist==1.4.0
func-timeout==4.3.5
grpcio==1.51.1
grpcio-tools==1.44.0
httplib2==0.20.2
idna==3.3
importlib-metadata==4.6.4
incremental==21.3.0
isabelle-client==0.4.1
jeepney==0.7.1
Jinja2==3.0.3
joblib==1.3.2
jsonpatch==1.32
jsonpointer==2.0
jsonschema==3.2.0
keyring==23.5.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
MarkupSafe==2.0.1
more-itertools==8.10.0
multidict==4.7.6
munkres==1.1.4
nest-asyncio==1.5.7
netifaces==0.11.0
networkx==3.2.1
numpy==1.25.2
oauthlib==3.2.0
openai==0.27.8
pexpect==4.8.0
prettytable==2.5.0
protobuf==3.20.3
psutil==5.9.5
ptyprocess==0.7.0
py4j==0.10.9.7
pyasn1==0.4.8
pyasn1-modules==0.2.1
PyHamcrest==2.0.2
PyJWT==2.3.0
pyOpenSSL==21.0.0
pyparsing==2.4.7
pyrsistent==0.18.1
pyserial==3.5
pyspark==3.5.0
python-magic==0.4.24
pytz==2022.1
PyYAML==6.0.1
requests==2.31.0
SecretStorage==3.3.1
service-identity==18.1.0
six==1.16.0
ssh-import-id==5.11
submitit==1.4.5
timeout-decorator==0.5.0
tqdm==4.65.0
Twisted==22.1.0
typing_extensions==4.7.1
unattended-upgrades==0.1
urllib3==1.26.5
wadllib==1.3.6
wcwidth==0.2.5
yarl==1.9.2
z3-solver==4.12.2.0
zipp==1.0.0
zope.interface==5.4.0
